+++ mythtv/libs/libmythtv/mythplayer.cpp
@@ -845,7 +845,7 @@ void MythPlayer::SetVideoParams(int width, int height, double fps,
             float temp_speed = (play_speed == 0.0f) ?
                 audio.GetStretchFactor() : play_speed;
             SetFrameInterval(kScan_Progressive,
-                             1.0 / (video_frame_rate * temp_speed));
+                             1.0 / (video_frame_rate * static_cast<double>(temp_speed)));
         }
     }
 
@@ -877,7 +877,7 @@ void MythPlayer::SetFrameRate(double fps)
     float temp_speed = (play_speed == 0.0f) ?
         audio.GetStretchFactor() : play_speed;
     SetFrameInterval(kScan_Progressive,
-                        1.0 / (video_frame_rate * temp_speed));
+                     1.0 / (video_frame_rate * static_cast<double>(temp_speed)));
 }
 
 void MythPlayer::SetFileLength(int total, int frames)
@@ -1847,7 +1847,8 @@ void MythPlayer::InitAVSync(void)
                        .arg(1000000.0 / frame_interval, 0, 'f', 3);
         LOG(VB_PLAYBACK, LOG_INFO, LOC + msg);
 
-        SetFrameInterval(m_scan, 1.0 / (video_frame_rate * play_speed));
+        SetFrameInterval(m_scan,
+                         1.0 / (video_frame_rate * static_cast<double>(play_speed)));
 
         // try to get preferential scheduling, but ignore if we fail to.
         myth_nice(-19);
@@ -2742,7 +2743,7 @@ void MythPlayer::VideoStart(void)
     refreshrate = frame_interval;
 
     float temp_speed = (play_speed == 0.0f) ? audio.GetStretchFactor() : play_speed;
-    int fr_int = (1000000.0 / video_frame_rate / temp_speed);
+    int fr_int = (1000000.0 / video_frame_rate / static_cast<double>(temp_speed));
     int rf_int = MythDisplay::GetDisplayInfo(fr_int).Rate();
 
     // Default to Interlaced playback to allocate the deinterlacer structures
@@ -4037,7 +4038,7 @@ bool MythPlayer::UpdateFFRewSkip(void)
         skip_changed = (ffrew_skip != 1);
         if (decoder)
             m_fpsMultiplier = decoder->GetfpsMultiplier();
-        frame_interval = (int) (1000000.0f / video_frame_rate / temp_speed)
+        frame_interval = (int) (1000000.0 / video_frame_rate / static_cast<double>(temp_speed))
           / m_fpsMultiplier;
         ffrew_skip = (play_speed != 0.0f);
     }
@@ -4055,7 +4056,7 @@ bool MythPlayer::UpdateFFRewSkip(void)
         frame_interval = (fabs(play_speed) >=  60.0f) ? 133466 : frame_interval;
         frame_interval = (fabs(play_speed) >= 120.0f) ? 133466 : frame_interval;
         frame_interval = (fabs(play_speed) >= 180.0f) ? 133466 : frame_interval;
-        float ffw_fps = fabs(play_speed) * video_frame_rate;
+        float ffw_fps = fabs(static_cast<double>(play_speed)) * video_frame_rate;
         float dis_fps = 1000000.0f / frame_interval;
         ffrew_skip = (int)ceil(ffw_fps / dis_fps);
         ffrew_skip = play_speed < 0.0f ? -ffrew_skip : ffrew_skip;
@@ -5279,7 +5280,7 @@ int MythPlayer::GetSecondsBehind(void) const
     if (played < 0)
         played = 0;
 
-    return (int)((float)(written - played) / video_frame_rate);
+    return (int)(static_cast<double>(written - played) / video_frame_rate);
 }
 
 int64_t MythPlayer::GetSecondsPlayed(bool honorCutList, int divisor) const
